git-svn-id: http://talendforge.org/svn/tom/branches/branch-5_1@92306 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.core/src/com/amalto/core/save/context/ChangeReferenceTypeAction.java
@@ -32,6 +32,10 @@ class ChangeReferenceTypeAction extends AbstractChangeTypeAction {
     }
 
     public MutableDocument perform(MutableDocument document) {
+        if (!hasChangedType) {
+            document.createAccessor(path).touch();
+            return document;
+        }
         // Ensure tmdm prefix is declared
         Document domDocument = document.asDOM();
         String xsi = domDocument.lookupNamespaceURI(SkipAttributeDocumentBuilder.TALEND_NAMESPACE); //$NON-NLS-1$
@@ -53,13 +57,21 @@ class ChangeReferenceTypeAction extends AbstractChangeTypeAction {
     }
 
     public MutableDocument undo(MutableDocument document) {
+        if (!hasChangedType) {
+            document.createAccessor(path).touch();
+            return document;
+        }
         Accessor accessor = document.createAccessor(path + "/@tmdm:type"); //$NON-NLS-1$
         accessor.delete();
         return document;
     }
 
     public String getDetails() {
-        return "Change FK type to " + newType.getName(); //$NON-NLS-1$
+        if (hasChangedType) {
+            return "Change FK type to " + newType.getName(); //$NON-NLS-1$
+        } else {
+            return "Change FK type to " + newType.getName() + " (NO OP)"; //$NON-NLS-1$ //$NON-NLS-2$
+        }
     }
 
     @Override
@@ -67,6 +79,7 @@ class ChangeReferenceTypeAction extends AbstractChangeTypeAction {
         return "ChangeReferenceTypeAction{" + //$NON-NLS-1$
                 "path='" + path + '\'' + //$NON-NLS-1$
                 ", newType=" + newType + //$NON-NLS-1$
+                ", changedType= " + hasChangedType + //$NON-NLS-1$
                 '}';
     }
